resource/cloudflare_page_rule: Swap `cache_key_fields` to use TypeSet
When this resource was introduced, it was done so using `TypeList` for
the inner structure. In Terraform, `TypeList` is used for ordered lists
whereas `TypeSet` is for unordered.

This was working for the most part however the `cache_key_fields` fields
don't place any importance on the order so they are returned from the
API with inconsistent ordering.

To address this issue, we need to swap the representation to be
`TypeSet` and ignore any changes in the ordering when comparing to the
state.

Fixes #890