provider: Expose integration test variables
In tests and payloads, we quite often use randomly generated IDs. This
is fine however it does lead to quite a bit of repetition or hardcoded
values which are a pain to change and can hide subtle bugs. Instead,
I've updated the provider test to expose the same details we set in the
`env` for CI and can reuse wherever we need to compare the values.

This means that in the future we can trim our tests down to reuse these
values without confusion.