test/cloudflare_magic_firewall_ruleset: skip CI tests for default zone
Like BYOIP, Magic Transit doesn't really lend itself to the typical CI
setup where we create and destroy resources on the fly. As such, we need
to update the test suite to work accordingly. The tests are very
valuable and we want them around however we shouldn't run them on a zone
we know isn't configured (in this case, able to be configured to do so).

This updates the acceptance tests to skip when the default zone ID is in
use however will allow everyone else to run the tests like normal.

Fixes https://github.com/cloudflare/terraform-provider-cloudflare/runs/1810568088?check_suite_focus=true#step:6:317