cloudflare_page_rule: handle lists and string values for `cache_key_fields`
The page rule API for defining cache keys allows you to include/exclude
either a list or all of the query parameters. However, to apply the
changes to all, the value sent must be a string, not a list as we've
previously sent (due to reusing `[]interface{}{"*"}`). In the case of
sending `exclude = ["*"]` it excludes the "*" parameter from the
cache key, not all of the possible keys.

Closes #967