pull out v0 => v1 cloudflare_record test case
this schema migration was added in 2014 for migrating Cloudflare records
from the old format to align with the V4 API and it has an assumption
that only a single request will be made in sequence. This is no longer
true and spending the time on fixing the test is a waste of time given
no one should be upgrading from a 7 year old schema.