resource/cloudflare_record: support `allow_overwrite`
There are scenarios, such as AWS ACM, which provide a DNS record for
ownership validation and due to these validation strings not being
uniquem users can hit problems with duplicated DNS records even though
the validation string remains the same.

While this **isn't** the norm for most it is painful for some so to
alleviate that, we've introduce the `allow_overwrite` argument which
will allow users to disregard certain record values and force their own
regardless of whether a record exists in that space or not. This will
allow modules that handle this on behalf of users to operate freely
without manual intervention.

Closes #960