resource/cloudflare_ruleset: fix state handler
While looking into upgrading to Terraform SDK v2, I found that the state file
was throwing some errors when trying to save `rules`. It turns out that SDK v1
was more forgiving with the types when it came to flattening lists however SDK
v2 is much stricter causing the failures. This fixes issue by correctly
building the state struct and flattening out some of the nested structures.