resource/cloudflare_ruleset: conditionally set action parameter "version"
Updates the action parameter generation process to only set the
"version" value when it is supplied by the end user. This addresses
issues where it was sent unconditionally and prevented some phases from
applying correctly. If the value is set in the configuration, the value
is still sent however, that will be a user configuration error, not a
provider error.

Closes #1309