[resource/device_posture_integration] Remove extra id parameters
the terraform-plugin-sdk adds the "id" to the schema when d.SetID is
used. therefore, there is no need to have this extra parameter in
the schema.

The use of this extra id is causing the pulumi schema to pull a
duplicate argument

```
sdk/python/pulumi_cloudflare/device_posture_integration.py:377: error: Duplicate argument "id" in function definition
```