resource/cloudflare_zone: don't get stuck in a loop for partner plans
In #1532 we added support for partner rate plans based on the _subscription_
endpoints however, within zone creation we get the _zone_ endpoint (same thing,
different "view of the world").

This caused an issue where we were waiting for a plan ID to return that never
would as the partners rate plans in the zone endpoint just display the regular
rate plan instead of the partner specific one.

Closes #1532