provider: allow configuring `x-auth-service-key` individually
Following in the footsteps of
https://github.com/cloudflare/terraform-provider-cloudflare/issues/1768#issuecomment-1182745202,
this introduces the ability to set the x-auth-service-key individually without
setting another authentication mechanism. Prior to 3.18, you could set multiple
and the internals looked busted but sometimes worked...and other times didn't.

This is hopefully the last case we need to tidy up to correctly set the auth
correctly for the request that users are making.

Closes #1905