docs: remove prefixed files for data-sources
When I fixed the data-source template back in #1899, I inadvertently "fixed" the
generation of the other resources too. This created duplicates in the data
source docs because the old website format required the full name
`cloudflare_zones` as opposed to `zones` and the documentation generation
copies over the older style verbatim.

Closes #1927