resource/cloudflare_access_rule: remove ability to create user level rules
Prior to this commit, if you omitted `account_id` and `zone_id` it would create
a user level access rule. This wasn't noted anywhere and caught a lot of people
out prior to releasing the WAF.

Here, we are removing that and explicitly requiring a zone or account ID as in
most cases, user level rules are the same as account level rules -- just have a
different name externally.