resource/cloudflare_certificate_pack: handle deletion via UI
The deletion policy for certificate packs is a little different and for
a short period of time following deletion, the URL does not return a 404
like most other resources. Instead, it will return a HTTP 200 with the
`status` field set to deleted. This isn't standard handling so we'll
explicitly account for both of these scenarios.