resource/cloudflare_ruleset: handle `Import` operations with missing values
In the event a user provides an invalid import string, we should provide
a nice error message stating which part is missing. Unfortunately, we
were attempting to perform assignments on those index parts which, when
they are missing, will cause a panic. This moves the assignment to under
the checks and uses the raw indexes for the values _if_ required.