Update preservation of rules to use refs rather than IDs
This is safe because all rules have a ref, as it defaults to the ID if
it is not explicitly set. The advantage of using refs is that we can
take into account whether the user has explicitly set the ref
themselves. If they have, then we avoid changing it, or using that ref
for other rules.