Make some minor fixes to the Rulesets acceptance resource tests
Use the `skip` action, rather than the `allow` action. The `allow`
action does not exist in Rulesets, and only existed in Firewall Rules.

Make the Origin Rules tests agnostic of the zone they are run with.

Add `(cf.zone.plan eq "ENT")` to the expression of account-level WAF
rulesets, as this is a new requirement.

Also add the `stringplanmodifier.UseStateForUnknown()` modifier to the
description attribute for rules, since this should never change when
it is computed.