push `empty` logic into `flattenTeamsRuleSettings`
To ensure we always set the `rules_setting` values (even when we are aiming to
remove the setting) I've moved the `empty` logic into the
`flattenTeamsRuleSettings` method so that the Read/Create logic are using the
same source. This also fixes a logic bug in the comparison where a `!= nil`
should have been a `== nil`.