Set the default value for the cache_key.exclude_origin setting to false
The API response by default will return a value for false for the exclude_origin value in the custom cache key setting. This adjusts the default value for the schema to be false to account for this behavior from the API. An extension to the tests was added to ensure backwards compatibility in cases where exclude_origin was not set for the cache key mechanisms.