resource/cloudflare_worker_script: handle creation and cleanup of R2 buckets in bindings
Updates the Worker tests to programmatically call the S3 compatible APIs
for R2 to create and clean up buckets since we cannot mix V5 and V6
protocols in the provider without circular dependencies. Once that is
resolved, we should swap to the inbuilt resource instead of these extra
API calls.