provider: add support for UA operator suffixes
Within the Terraform ecosystem, there are multiple ways of using the
providers. Generally, it is through the Terraform CLI however, there are
third party tools like Pulumi that rely on the schemas to build their
own integrations.

One issue with this is that the requests they make are lumped with the
label of `terraform/<version>` in the HTTP user agent which doesn’t
allow observablility into those specific requests. To enable better
visibility here, we are introducing a new schema field called a User
Agent Operator suffix. It is configurable in the schema
(`user_agent_operator_suffix`) and via the environment variable
(`CLOUDFLARE_USER_AGENT_OPERATOR_SUFFIX `) and will provide a way for
these third party integrations to denote the source of their requests.

⚠️This *is not* a recommended configuration change for most and when not
explicitly configured will default to the existing `terraform/<version>`
string (as it always has). You should only change this value if you know
what you are doing.

Signed-off-by: Jacob Bednarz <jacob.bednarz@gmail.com>