provider: introduce `FindGoModuleVersion` helper for module versions
Before this change, we relied on the plugin libraries to expose their version. Unfortunately, this was incorrectly set in SDKv2[1] and non-existent in the framework[2].

The recommended approach is to instead dig into the build information via`runtime/debug.ReadBuildInfo()`[3] which is what we introduce here in a slightly safer and consistent way.

[1]: https://github.com/hashicorp/terraform-plugin-sdk/issues/1257
[2]: https://github.com/hashicorp/terraform-plugin-framework/issues/855
[3]: https://pkg.go.dev/runtime/debug#ReadBuildInfo

Signed-off-by: Jacob Bednarz <jacob.bednarz@gmail.com>