provider: introduce a muxed client
With the announcement of the cloudflare-go/v2[1], we need to introduce a
way to use both versions of the SDK in the provider. This moves us to a
world where the v1 and v2 of the clients can coexist and still be
initialised in a consistent fashion.

At a high level, the guidance for modifying resources and datasources is:

- If you have an existing resource, use the cloudflare-go V1 SDK which
  whatever version of the Terraform SDK/framework version you have
  today.
- If you have a new resource, build it using cloudflare-go V2 SDK and
  the terraform-plugin-framework.

While this commit is rather large by LOC, it is the following changes
but made across the provider.

- Introduce a `muxclient` package. Allows Terraform internals to have a
  single client instead of trying to swap on the fly potentially causing
  concurrency issues.
- Split `acctest.SharedClient` into `SharedV1Client` and `SharedV2Client`
  for use in both versions of acceptance tests.
- Aliases and namespacing to handle the different package versions.

These changes should be internal and not end user facing.
[1]: https://blog.cloudflare.com/workers-production-safety