resource/cloudflare_list_item: handle overlapping redirect `source_url`
To determine the individual list item identifier, we perform an
additional HTTP search against the endpoint filtering for the value. In
the case of IPs, we use a "starts with" match whereas for redirects, it
is a contains. This creates an edge whereby if you have a multiple
redirects whereby they share a substring, you cannot differentiate them.

This introduces an additional loop performing an exact match against
what the user has provided with all found search results.

NB: For a small dataset, this is manageable however, for 10+ similar
items, you are very likely to exhaust your API quota due to additional
requests. In the future, the service will improve the search to provide
different search match types which we can lean on here instead.