resource/cloudflare_zone_override: don't rip out schemas...yet
Updates the handling of the deprecation to be a little more user friendly given
the API is hanging out, but will always return the "off" status.

The properties will remain around with a state migration taking care of clearing
out the values (although, kinda redundant but 🤷‍♀️). Leaving the configuration
file cleanup to the end user as the only step required here.

Then in the next major release, we'll remove the schemas entirely.