Fix ZT lists update
When updating the `cloudflare_zero_trust_list` only changes to the `items`
field was being considered to determine the list elements to store.
However, customers can also use the `items_with_description` field to
declare list elements with a description, so it also needs to be
considered. This commit fixes the issue, adding an acceptance test to
prove its effectiveness.