dackerman: sync with generated
Somehow the next branch got out of sync with generated, and had reverted
some error handling, which was swallowing the error when encoding the
nested types.

This commit makes the file match generated so there is no longer a custom
code diff, and that means going forward you'll see an error message if
your custom types don't support "unknown" or "null" and therefore aren't
encoding properly.