Remove logic to preserve ruleset rule refs
This logic tries to detect which rules are the same as before during an
update, and preserve the refs (and therefore IDs) of these rules.

However, this logic is fallible, since we cannot know the user's true
intention, and makes it difficult for us to remove the `version` and
`last_updated` fields from rules.

If users want to preserve rule refs across updates, they can now
actually use the `ref` field directly—and there is no need for this
remapping logic anymore.