Remove `version` and `last_updated` fields from ruleset rules
These fields frequently change, leading to verbose diffs in plans. Since
these fields are not really useful within Terraform, we remove them from
the schema completely to prevent them from "polluting" diffs.

This also matches the behavior we have for the ruleset-level `version`
and `last_updated` fields, which are also not represented in the
Terraform schema.