fix(api_token): store `value` out of band of marshaler
For API tokens, `value` is a write only property that is only present in
the POST response when creating a new one. Outside of that response,
it is not accessible (by design).

To work around that, we need to merge the marshaled response from the
GET upon refresh. This ensures you can still reference the value and it
doesn't get removed from the state.