fix(dns_record): relax constraint for overlapping unions
When the field is a part of a union that overlaps with other
configurations, we currently take the first instance and apply the
constraints from that to the field. This kinda works but gets difficult
to automatically resolve so instead, let's remove this and instead rely
on the remote to do the stricter validation.