fix(workers_script): fix incorrect resource refreshing
Previously, the provider was hitting an endpoint that returned solely the script's content rather than any resource data, resulting in the refreshed state being completely null. Now, the provider will fetch the up-to-date resource data from the handful of endpoints it needs to in order to piece together the fresh state. In a future, we hope to return a complete representation of the resource from a single endpoint.