TUN-9171: Fix vnet resource by mapping IsDefaultNetwork to IsDefault
## Summary
The virtual network resource uses a different field  to set and to read the same value. Therefore, to have the resource properly
working on terraform v5, we need to map the value read to the value set. For this case the value used to write is `isDefault` and the value read, is `isDefaultNetwork`.