fix(workers_script): remove `UseStateForUnknown` plan modifier on `startup_time_ms` attribute
This plan modifier was originally added to reduce noisy plan diffs (#5544), but it turns out the value returned by the API is not static. This resulted in Terraform sometimes raising an error that the resulting state didn't match planned state (because this value was no longer treated as unknown).

Fixes #5704