DS-15222: Add LogpushJob_Full test to test all fields
This adds `LogpushJob_Full` test to test all fields.

This is to demonstrate what changes need to be made on `model.go` and `schema.go`, which are manually done here to make the test passes.

Notable changes are the following:
 1. Remove `Computed` and `Default` from some fields to match with Logpush API.
 2. Added `Computed` for `Enabled` field, as Logpush API always return value (default `false`).
 2. Change `max_upload_*` to be integer, with range validator.
 3. Adding `no_refresh` attribute for the fields that cause "drift" due to Logpush API json `omitempty`.

Some changes would need to be applied to Logpush API OpenAPI schema to make same changes on the next SDK generation.