fix(workers_script): fix nested computed attributes under `bindings` causing unwarranted plan diffs
Terraform lacks support for resource schemas to have polymorphic collection types (like `bindings`, whose element type can be thought of as a union of types discriminated by their `type` attribute). This means that from Terraform's perspective, all bindings are of one large (undiscriminated) type containing the attributes of each binding. This also means that if some binding type has an attribute that is Computed in the schema, then that attribute will be "unknown" at plan time (assuming it's not configured) for all bindings, even if their binding type doesn't include that attribute. Worse yet, if some binding attribute is both Computed and Optional, Terraform will plan to update the resource if there are any bindings at all that leave that attribute unconfigured (which is very likely given most binding types don't include that attribute in reality). This was causing `namespace_id = (known after apply)` to appear during plans even when a resource hadn't been changed.

This commit introduces a new plan modifier `UnknownOnlyIf` to work around this lack of support for discriminated types. Here, we're telling Terraform to treat these computed attributes as unknown ONLY IF this binding's `type` is `durable_object_namespace`. Otherwise, revert the plan back to null for this attribute.