Fix regressions in drift prevention `cloudflare_ruleset`
Previously, there was a bug where `state.Rules` was being used instead
of `plan.Rules` in the `ModifyPlan` function, which is designed to
prevent drift.

Additionally, the original plan was never updated in `ModifyPlan`,
resulting in changes to the plan to never be applied.

Separately, rule descriptions were incorrectly marked as computed and as
having a default of `""`, causing drift when a description was not set.

* Fix the regressions in the `ModifyPlan` function, also simplifying the
  code to use `.AsStructSliceT()` rather than `.Elements()`.

* Mark rule descriptions as non-computed and remove their default value.
  The OpenAPI schema has already been updated to match this, so this
  change will match the next Stainless update.

* Add plan check tests to prevent future regressions of the drift
  prevention logic.