don't treat `bindings.*.version_id` as Computed and remove default value (#6249)
The `version_id` binding property is only relevant to bindings of type `inherit`, which allow a user to specify a binding whose value should be copied from another version. If `version_id` is unspecified on an inherit binding, it's assumed to be `latest` by default As a result, the attribute was codegen'd to be Computed and have a `Default` value.

However, `inherit` bindings are only valid in the context of a request and will never be returned by the API. Thus, it doesn't make sense for this attribute to be Computed (or have a default).